% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idm.R
\name{create_constant_hazard}
\alias{create_constant_hazard}
\title{Create constant hazards}
\usage{
create_constant_hazard(a12_const, a13_const, a23_const)
}
\arguments{
\item{a12_const}{Numeric. Constant hazard for 1→2 (healthy to illness) transition.}

\item{a13_const}{Numeric. Constant hazard for 1→3 (healthy to death) transition.}

\item{a23_const}{Numeric. Constant hazard for 2→3 (illness to death) transition.}
}
\value{
idm_estimators object.
}
\description{
Creates constant hazard functions for the illness-death model.
}
