% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idm.R
\name{create_weibull_hazard}
\alias{create_weibull_hazard}
\title{Create Weibull hazards}
\usage{
create_weibull_hazard(shape12, scale12, shape13, scale13, shape23, scale23)
}
\arguments{
\item{shape12}{Numeric. Shape parameter for Weibull hazard from state 1 to 2.}

\item{scale12}{Numeric. Scale parameter for Weibull hazard from state 1 to 2.}

\item{shape13}{Numeric. Shape parameter for Weibull hazard from state 1 to 3.}

\item{scale13}{Numeric. Scale parameter for Weibull hazard from state 1 to 3.}

\item{shape23}{Numeric. Shape parameter for Weibull hazard from state 2 to 3.}

\item{scale23}{Numeric. Scale parameter for Weibull hazard from state 2 to 3.}
}
\value{
idm_estimators object.
}
\description{
Create Weibull hazards
}
