% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmEstimation-package.R
\docType{package}
\name{idmEstimation-package}
\alias{idmEstimation}
\alias{idmEstimation-package}
\title{idmEstimation}
\description{
Tools for IDM estimation (write a better description here).
}
\author{
\strong{Maintainer}: Rasmus Skousen \email{rasmus.emils@proton.com}

}
\keyword{internal}
