% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idm.R
\name{simulate_idm_frydman}
\alias{simulate_idm_frydman}
\title{Simulate Illness-Death Data with using Frydman and Szareks's approach}
\usage{
simulate_idm_frydman(n, scenario = 1L)
}
\arguments{
\item{n}{Integer. Number of subjects to simulate. Default is 1000.}

\item{scenario}{Determining censoring scheme default is 1.}
}
\value{
An object of class \code{c("idm_object", "idm_exact_object")} containing:
\describe{
\item{data}{Simulated observed data frame}
\item{model_type}{Character string describing the simulation model}
\item{model_config}{List with simulation parameters and summaries}
\item{estimators}{True hazard, cumulative hazard, and distribution functions}
\item{exact_idm}{Data frame with true (uncensored) event times}
}
}
\description{
Simulate Illness-Death Data with using Frydman and Szareks's approach
}
